// src/particles/themeInterface.d.ts
export interface ITheme {
    name: string,
    colors: {
        bg: string;
        bgResultField: string;
        textContrast: string;
        buttonText: string;
        resultText: string;
        bgControlsContainer: string;
        bgButtonNumber: string;
        bgSwitchButton: string;
        bgButtonAlt: string;
        bgButtonResult: string;
        shadowButtonAlt: string;
        shadowResult: string;
        shadowButton: string;
    }
}